Revised the broker implementation of bulk-requests, fixing bugs and adding a single new feature so that it correctly displays an ontology-specific bulk request file.
(1) Fixed two bugs in BulkTable, so that lines were correctly added (and so that their metadata fields were correctly mapped to column headings). 
(2) Fixed bugs, in LocalBroker, in the way that User and Ontology IDs were parsed.
(3) Changed the Broker interface, so that listRequests() now takes an Ontology argument (instead of a String ontology_id).
(4) Updated Ontology with hashCode() and isEqual() methods.
(5) Added a redirect to the NewBulkRequestServlet, so redirect .../ontology/#/bulk-requests to the /bulk-requests?ontology_id=[URL] where [URL] is the url of the original ontology requested.  This is a handy way to do ontology-specific bulk requests.
(6) Updated index.html, to ask for bulk-requests from ontology 1 in the static link at the bottom.  (This is a hack, and should be changed.) 

git-svn-id: http://svn.neurocommons.org/svn/trunk/broker@3834 301305c5-f827-0410-9252-dba5032f77be